% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_gnucash.R, R/ingest_quickbooks.R
\name{ingest_gnucash_xml}
\alias{ingest_gnucash_xml}
\title{Ingest GNUCash XML file
Note yet implemented!}
\usage{
ingest_gnucash_xml(path, practice_id, account_map = default_account_map())

ingest_gnucash_xml(path, practice_id, account_map = default_account_map())
}
\description{
Ingest GNUCash XML file
Note yet implemented!

Ingest Quickbooks file
Note yet implemented!
}
